+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `opentelemetry-instrumentation-asgi` now returns a `traceresponse` response header.
   ([#817](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/817))
+- `opentelemetry-instrumentation-kafka-python` added kafka-python module instrumentation.
+  ([#814](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/814))
 
 ### Fixed